se inicia la creación de la lista de items en pantalla

+++ domain/src/main/java/com/alexis/domain/model/ValueStruct.java
@@ -1,6 +1,6 @@
 package com.alexis.domain.model;
 
 public class ValueStruct {
-    public int number;
+    public double number;
     public String unit;
 }
